requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
lxml==4.9.3
python-dateutil==2.8.2
urllib3==2.1.0
chardet==5.2.0
numpy==1.24.4
openpyxl==3.1.2
python-dotenv==1.0.0
openai>=1.0.0